requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
certifi==2023.11.17
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
magic-filter==1.0.12
multidict==6.0.4
openai==1.7.2
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
yarl==1.9.4

SQLAlchemy~=2.0.25